Made it so that people can't edit/delete other people's favors

+++ barter/views.py
@@ -139,6 +139,9 @@ def user_edit(request, pk):
 
 @login_required
 def favor_edit(request, pk):
+    if int(request.user.pk) is not int(pk):
+        messages.error(request, 'You are not the author of this favor.')
+        return HttpResponseRedirect("/favors/" + pk)
     favor = Favor.objects.get(pk=pk)
     form = FavorEditForm(request.POST or None, initial={'title': favor.title, 'message': favor.message})
     if form.is_valid():
@@ -152,6 +155,9 @@ def favor_edit(request, pk):
 
 @login_required
 def favor_delete(request, pk):
+    if int(request.user.pk) is not int(pk):
+        messages.error(request, 'You are not the author of this favor.')
+        return HttpResponseRedirect("/favors/" + pk)
     favor = Favor.objects.get(pk=pk)
     favor.delete()
     messages.success(request, 'Favor deleted.')
